# _archive/phase0_baseline/openapi_after_phase_a.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise SDN Controller \u2014 Core Ingress & Validation Orchestrator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v5/discovery/on-boarding-ingestion": {
      "post": {
        "tags": [
          "ZTP Discovery"
        ],
        "summary": "Ingest Ztp Signal",
        "description": "Ingest a ZTP discovery signal from a newly unboxed switch.",
        "operationId": "ingest_ztp_signal_api_v5_discovery_on_boarding_ingestion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZtpIngestionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/discovery/pool": {
      "get": {
        "tags": [
          "ZTP Discovery"
        ],
        "summary": "Get Discovery Pool",
        "description": "Get the onboarding queue status (pending/provisioned/failed) per entry.\nTenant-scoped: non-platform-admins only see switches in their tenant's fabrics.",
        "operationId": "get_discovery_pool_api_v5_discovery_pool_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v5/discovery/pool/{discovery_id}/status": {
      "get": {
        "tags": [
          "ZTP Discovery"
        ],
        "summary": "Get Ztp Record Status",
        "description": "Get full status details for a single ZTP discovery record.",
        "operationId": "get_ztp_record_status_api_v5_discovery_pool__discovery_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "discovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Discovery Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/discovery/pool/{discovery_id}/retry": {
      "post": {
        "tags": [
          "ZTP Discovery"
        ],
        "summary": "Retry Ztp Provisioning",
        "description": "Retry provisioning for a failed ZTP discovery record.",
        "operationId": "retry_ztp_provisioning_api_v5_discovery_pool__discovery_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "discovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Discovery Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/discovery/pool/{discovery_id}/assign-fabric": {
      "patch": {
        "tags": [
          "ZTP Discovery"
        ],
        "summary": "Assign Switch To Fabric",
        "description": "Assign a newly-discovered switch to a fabric.\nThis dynamically calculates and allocates:\n- Next available BGP ASN from the fabric global ASN.\n- Loopback IP from fabric.loopback_pool.\n- VTEP IP from fabric.vtep_pool.\nAutomatically re-triggers the ZTP baseline provisioning task.",
        "operationId": "assign_switch_to_fabric_api_v5_discovery_pool__discovery_id__assign_fabric_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "discovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Discovery Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignFabricPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v5/discovery/pool/{discovery_id}": {
      "delete": {
        "tags": [
          "ZTP Discovery"
        ],
        "summary": "Remove Ztp Record",
        "description": "Remove a ZTP discovery record. Platform Admin only.",
        "operationId": "remove_ztp_record_api_v5_discovery_pool__discovery_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "discovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Discovery Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignFabricPayload": {
        "properties": {
          "fabric_id": {
            "type": "string",
            "title": "Fabric Id",
            "description": "UUID of the fabric to assign this switch to"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role",
            "description": "Switch role: 'spine' or 'leaf'",
            "default": "leaf"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname",
            "description": "Override hostname (leave blank to keep current)"
          }
        },
        "type": "object",
        "required": [
          "fabric_id"
        ],
        "title": "AssignFabricPayload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ZtpIngestionPayload": {
        "properties": {
          "mac_address": {
            "type": "string",
            "title": "Mac Address",
            "description": "Chassis base interface physical MAC address identifier"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number",
            "description": "Switch chassis physical hardware serial identifier"
          },
          "os_version": {
            "type": "string",
            "title": "Os Version",
            "description": "Active operating system core release string"
          },
          "vendor": {
            "type": "string",
            "title": "Vendor",
            "description": "Vendor target flag matching supported drivers"
          },
          "management_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Management Ip",
            "description": "Optional IP override for testing/simulation"
          }
        },
        "type": "object",
        "required": [
          "mac_address",
          "serial_number",
          "os_version",
          "vendor"
        ],
        "title": "ZtpIngestionPayload"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}